backend/requirements.txt
# Core Python
python-dotenv==1.1.0
pydantic==2.11.4
pydantic-settings==2.9.1

# Web Framework
fastapi==0.115.12
uvicorn==0.34.2
slowapi==0.1.9
python-multipart==0.0.20
httpx==0.28.1

# Database
motor==3.7.0
pymongo==4.12.1
beanie==1.29.0
SQLAlchemy==2.0.40
pymysql
redis==6.0.0
sqlparse==0.5.3

# Authentication & Security
python-jose==3.4.0
passlib==1.7.4
bcrypt==4.3.0
pyotp==2.9.0
email-validator==2.2.0

# AI & LLM
langgraph
langwatch==0.2.0
langchain==0.3.25
langchain-core==0.3.60
langchain-google-genai==2.1.4
langchain-tavily==0.1.6
dspy==2.6.24

# Browser Automation
playwright==1.52.0
browser-use==0.1.39

# Utilities
websockets==15.0.1
aiofiles==24.1.0

# Testing
pytest==8.3.5